- patches.drivers/libata-acpi-size_t-fix: Fix printk format   with size_t. - patches.fixes/fix-compare-with-sizeof: [PATCH trivial] Fix   the comparison with sizeof(). - patches.fixes/fix-kauditd_thread-return-value: [PATCH trivial]   Add missing return value of kauditd_thread. - patches.fixes/jffs2-size_t-fix: [PATCH] jffs2: Fix size_t on   64bit architectures. - patches.fixes/scsi-missing-iomem-cast: [PATCH] Add cast to   __iomem pointer in scsi drivers. - patches.fixes/value_computed_not_used-fix: [PATCH trivial]   Fix "value computed not used" warnings. - patches.suse/lkcd-compile-warning-fix: Fix compile warning   with lkcd patch.
suse-commit: fac7ec1f09379813bd2c391086f2c490b150db98

+++ drivers/scsi/aacraid/aachba.c
@@ -1597,13 +1597,13 @@ int aac_scsi_cmd(struct scsi_cmnd * scsicmd)
 		cp[11] = 0;
 		cp[12] = 0;
 		aac_internal_transfer(scsicmd, cp, 0,
-		  min((unsigned int)scsicmd->cmnd[13], sizeof(cp)));
+		  min((size_t)scsicmd->cmnd[13], sizeof(cp)));
 		if (sizeof(cp) < scsicmd->cmnd[13]) {
 			unsigned int len, offset = sizeof(cp);
 
 			memset(cp, 0, offset);
 			do {
-				len = min(scsicmd->cmnd[13]-offset, sizeof(cp));
+				len = min((size_t)(scsicmd->cmnd[13]-offset), sizeof(cp));
 				aac_internal_transfer(scsicmd, cp, offset, len);
 			} while ((offset += len) < scsicmd->cmnd[13]);
 		}
